requirements.txt
# Core dependencies
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.1.0
slither-analyzer==0.10.0
langgraph==0.4.5
groq>=0.11.0
hedera-sdk-py>=2.19.0
reportlab==4.4.2
pygments==2.17.2
python-multipart==0.0.9
pydantic>=2.10.0

# HCS-10 integration
xxhash==3.5.0

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0
